Update 11-arrays-loops.html
Finally found the problem that kept my code in an endless loop.
I'd open the server in the browser and everything would hang and crash.
the error was in i = i++. (with time i'll understand and get to explain why this syntax kept my code in an endless loop).
For now, the solution was to change it to i++ which made the code finally run successfully.